+++ sdk/nodejs/docker.ts
@@ -184,7 +184,7 @@ async function buildAndPushImageWorkerAsync(
         throw new Error(`[repositoryUrl] should not contain a tag: ${repositoryUrlTag}`);
     }
 
-    const { imageName, tag } = utils.getImageNameAndTag(baseImageName);
+    const tag = utils.getImageNameAndTag(baseImageName).tag;
 
     // login immediately if we're going to have to actually communicate with a remote registry.
     //
@@ -219,11 +219,11 @@ async function buildAndPushImageWorkerAsync(
     if (pullFromCache) {
         const dockerBuild = <pulumi.UnwrappedObject<DockerBuild>>pathOrBuild;
         const cacheFromParam = typeof dockerBuild.cacheFrom === "boolean" ? {} : dockerBuild.cacheFrom;
-        cacheFrom = pullCacheAsync(imageName, cacheFromParam, repositoryUrl, logResource);
+        cacheFrom = pullCacheAsync(baseImageName, cacheFromParam, repositoryUrl, logResource);
     }
 
     // Next, build the image.
-    const { imageId, stages } = await buildImageAsync(imageName, pathOrBuild, logResource, cacheFrom);
+    const { imageId, stages } = await buildImageAsync(baseImageName, pathOrBuild, logResource, cacheFrom);
     if (imageId === undefined) {
         throw new Error("Internal error: docker build did not produce an imageId.");
     }
@@ -249,12 +249,12 @@ async function buildAndPushImageWorkerAsync(
         // nice and simple url that they can reach this image at, without having the explicit imageId
         // hash added to it.  Note: this location is not guaranteed to be idempotent.  For example,
         // pushes on other machines might overwrite that location.
-        await tagAndPushImageAsync(imageName, repositoryUrl, tag, imageId, logResource);
-        await tagAndPushImageAsync(imageName, repositoryUrl, tag, /*imageId:*/ undefined, logResource);
+        await tagAndPushImageAsync(baseImageName, repositoryUrl, tag, imageId, logResource);
+        await tagAndPushImageAsync(baseImageName, repositoryUrl, tag, /*imageId:*/ undefined, logResource);
 
         for (const stage of stages) {
             await tagAndPushImageAsync(
-                localStageImageName(imageName, stage), repositoryUrl, stage, /*imageId:*/ undefined, logResource);
+                localStageImageName(baseImageName, stage), repositoryUrl, stage, /*imageId:*/ undefined, logResource);
         }
     }